Interface - Issue with Large Records
So because the Record Query function returns a count of sets as an
ssize_t, it might be a really big number, too large for an int. This is
causing some issues with the error-check macros on really large records
over 2 GiB, since it loops around negative. Anyways, since large records
are pretty important for mass computation, I'm making this change.

+++ lib/iface.h
@@ -23,13 +23,13 @@
 
 #define CK_NO(NO) \
     do { \
-        int __no = (NO); \
+        long long __no = (NO); \
         if (__no > 0) FAULT(); \
     } while (0)
 
 #define CK_RES(RES) \
     do { \
-        int __res = (RES); \
+        long long __res = (RES); \
         if (__res < 0) FAULT(); \
     } while (0)
